Move read_trust_anchor() to be a method
