+++ crates/server/src/store/recursor/config.rs
@@ -21,10 +21,7 @@ use serde::Deserialize;
 
 use crate::error::ConfigError;
 #[cfg(feature = "__dnssec")]
-use crate::proto::{
-    dnssec::{TrustAnchor, Verifier},
-    serialize::txt::trust_anchor::{self, Entry},
-};
+use crate::proto::dnssec::TrustAnchor;
 use crate::proto::{
     rr::{Name, RData, Record, RecordSet},
     serialize::txt::Parser,
@@ -151,7 +148,7 @@ impl DnssecPolicyConfig {
             Self::ValidateWithStaticKey { path } => DnssecPolicy::ValidateWithStaticKey {
                 trust_anchor: path
                     .as_ref()
-                    .map(|path| read_trust_anchor(path))
+                    .map(|path| TrustAnchor::read_from_file(path))
                     .transpose()?
                     .map(Arc::new),
             },
@@ -159,45 +156,10 @@ impl DnssecPolicyConfig {
     }
 }
 
-#[cfg(feature = "__dnssec")]
-fn read_trust_anchor(path: &Path) -> Result<TrustAnchor, String> {
-    use std::fs;
-
-    let contents = fs::read_to_string(path).map_err(|e| e.to_string())?;
-
-    parse_trust_anchor(&contents)
-}
-
-#[cfg(feature = "__dnssec")]
-fn parse_trust_anchor(input: &str) -> Result<TrustAnchor, String> {
-    let parser = trust_anchor::Parser::new(input);
-    let entries = parser.parse().map_err(|e| e.to_string())?;
-
-    let mut trust_anchor = TrustAnchor::new();
-    for entry in entries {
-        if let Entry::DNSKEY(record) = entry {
-            let dnskey = record.data();
-            // XXX should we filter based on `dnskey.flags()`?
-            let key = dnskey.key().map_err(|e| e.to_string())?;
-            trust_anchor.insert_trust_anchor(&*key);
-        }
-    }
-
-    Ok(trust_anchor)
-}
-
 #[cfg(test)]
 mod tests {
+    #[cfg(all(feature = "__dnssec", feature = "toml"))]
     use super::*;
-
-    #[cfg(feature = "__dnssec")]
-    #[test]
-    fn can_load_trust_anchor_file() {
-        let input = include_str!("../../../../proto/tests/test-data/root.key");
-
-        let trust_anchor = parse_trust_anchor(input).unwrap();
-        assert_eq!(3, trust_anchor.len());
-    }
 
     #[cfg(all(feature = "__dnssec", feature = "toml"))]
     #[test]
